requirements.txt
amqpy==0.13.1
numpy<1.24.0,>=1.18.5
opencv-python==4.9.0.80
orjson==3.9.15
pandas==2.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
torchvision==0.17.1
requests==2.31.0
tqdm==4.66.2
pyyaml==6.0.1
matplotlib==3.8.3
seaborn==0.13.2
scipy==1.12.0
protobuf<4.21.3
tensorboard>=2.4.1
ipython==8.22.2
psutil==5.9.8
